# Optimize `set!` DSL